Big refactor for better tag handling, conciseness and DDC correctness
Previously, it was the responsibility of the instructions to update the
tag table, but now so long as they call the CHERI DDC/cap-based
loads/stores that interpose the MMU, this is all handled for them. DDC
was also unconditionally being added in the MMU, and no capability
checks performed on it. This is now done in the CHERI class when
instructions call the DDC-based methods, only needing to provide the
address as with the MMU. We also add PCC checks, add the missing LQ.CAP
instruction, delete the RV128-only LDU.CAP and update all the existing
loads/stores to have the correct XLEN checks, including making the
D-sized instructions be the capability loads/stores on RV32 for if Spike
ever gains support for CHERI-64.